Refactor AI and defense managers
