+++ src/game-systems/defense-systems/DefenseSystemManager.ts
@@ -1,5 +1,6 @@
 import type { Tower, TowerSlot, Position, Enemy, Effect } from '../../models/gameTypes';
 import { GAME_CONSTANTS } from '../../utils/constants';
+import { getDefenseRecommendations } from "./helpers/recommendations";
 
 /**
  * Defense System Manager
@@ -312,104 +313,11 @@ export class DefenseSystemManager {
       Math.pow(pos1.y - pos2.y, 2)
     );
   }
-
-  /**
-   * Get defense system recommendations
-   */
-  public getDefenseRecommendations(
-    towerSlots: TowerSlot[],
-    _enemies: Enemy[]
-  ): Array<{ 
-    type: 'shield' | 'repair'; 
-    position: Position; 
-    reason: string; 
-    priority: number 
-  }> {
-    const recommendations: Array<{ 
-      type: 'shield' | 'repair'; 
-      position: Position; 
-      reason: string; 
-      priority: number 
-    }> = [];
-
-    // Analyze tower clustering for shield placement
-    const towerClusters = this.findTowerClusters(towerSlots);
-    for (const cluster of towerClusters) {
-      if (cluster.towers.length >= 3) {
-        const centerPosition = this.calculateClusterCenter(cluster.towers);
-        recommendations.push({
-          type: 'shield',
-          position: centerPosition,
-          reason: `Protect ${cluster.towers.length} clustered towers`,
-          priority: cluster.towers.length * 10
-        });
-      }
-    }
-
-    // Analyze damaged towers for repair station placement
-    const damagedTowers = towerSlots.filter(slot => 
-      slot.tower && slot.tower.health < slot.tower.maxHealth * 0.8
-    );
-    
-    if (damagedTowers.length >= 2) {
-      const centerPosition = this.calculateClusterCenter(damagedTowers.map(s => s.tower!));
-      recommendations.push({
-        type: 'repair',
-        position: centerPosition,
-        reason: `Repair ${damagedTowers.length} damaged towers`,
-        priority: damagedTowers.length * 15
-      });
-    }
-
-    return recommendations.sort((a, b) => b.priority - a.priority);
+  public getDefenseRecommendations(towerSlots: TowerSlot[], enemies: Enemy[]) {
+    return getDefenseRecommendations(towerSlots, enemies);
   }
 
-  /**
-   * Find tower clusters for strategic placement
-   */
-  private findTowerClusters(towerSlots: TowerSlot[]): Array<{ towers: Tower[]; center: Position }> {
-    const clusters: Array<{ towers: Tower[]; center: Position }> = [];
-    const towers = towerSlots.filter(slot => slot.tower).map(slot => slot.tower!);
-    const visited = new Set<string>();
 
-    for (const tower of towers) {
-      if (visited.has(tower.id)) continue;
-
-      const cluster = this.findNearbyTowers(tower, towers, 150);
-      if (cluster.length >= 2) {
-        cluster.forEach(t => visited.add(t.id));
-        clusters.push({
-          towers: cluster,
-          center: this.calculateClusterCenter(cluster)
-        });
-      }
-    }
-
-    return clusters;
-  }
-
-  /**
-   * Find nearby towers within radius
-   */
-  private findNearbyTowers(centerTower: Tower, allTowers: Tower[], radius: number): Tower[] {
-    return allTowers.filter(tower => {
-      const distance = this.getDistance(centerTower.position, tower.position);
-      return distance <= radius;
-    });
-  }
-
-  /**
-   * Calculate center position of a cluster
-   */
-  private calculateClusterCenter(towers: Tower[]): Position {
-    const sumX = towers.reduce((sum, tower) => sum + tower.position.x, 0);
-    const sumY = towers.reduce((sum, tower) => sum + tower.position.y, 0);
-    
-    return {
-      x: sumX / towers.length,
-      y: sumY / towers.length
-    };
-  }
 }
 
 // Global defense system manager instance
